fix: prevent error when deleting empty entity

+++ Modules/CIPPCore/Public/Test-CIPPRerun.ps1
@@ -22,7 +22,9 @@ function Test-CIPPRerun {
         $RerunData = Get-CIPPAzDataTableEntity @RerunTable -filter "PartitionKey eq '$($TenantFilter)' and RowKey eq '$($Type)_$($API)'"
         if ($ClearAll.IsPresent) {
             $AllRerunData = Get-CIPPAzDataTableEntity @RerunTable
-            Remove-AzDataTableEntity @RerunTable -Entity $AllRerunData -Force
+            if ($AllRerunData) {
+                Remove-AzDataTableEntity @RerunTable -Entity $AllRerunData -Force
+            }
             return $false
         }
 
@@ -67,7 +69,7 @@ function Test-CIPPRerun {
     } catch {
         $ErrorMessage = Get-CippException -Exception $_
         Write-Host "Could not detect if this is a rerun: $($ErrorMessage.NormalizedError)"
-        Write-LogMessage -headers $Headers -API $API -message "Could not detect if this is a rerun: $($ErrorMessage.NormalizedError)" -Sev 'Error' -LogData $ErrorMessage
+        Write-LogMessage -headers $Headers -API $API -message "Could not detect if this is a rerun: $($ErrorMessage.NormalizedError)" -Sev 'Error' -LogData (Get-CippException -Exception $_)
         return $false
     }
 }
